+++ packages/gatsby-image/src/index.js
@@ -49,15 +49,51 @@ const convertProps = props => {
 }
 
 /**
+ * Checks if fluid or fixed are art-direction arrays.
+ *
+ * @param currentData  {{media?: string}[]}   The props to check for images.
+ * @return {boolean}
+ */
+const hasArtDirectionSupport = currentData =>
+  !!currentData &&
+  Array.isArray(currentData) &&
+  currentData.some(image => typeof image.media !== `undefined`)
+
+/**
+ * Tries to detect if a media query matches the current viewport.
+ * @property media   {{media?: string}}  A media query string.
+ * @return {boolean}
+ */
+const matchesMedia = ({ media }) =>
+  media ? isBrowser && !!window.matchMedia(media).matches : false
+
+/**
  * Find the source of an image to use as a key in the image cache.
  * Use `the first image in either `fixed` or `fluid`
- * @param {{fluid: {src: string}[], fixed: {src: string}[]}} args
+ * @param {{fluid: {src: string, media?: string}[], fixed: {src: string, media?: string}[]}} args
  * @return {string}
  */
 const getImageSrcKey = ({ fluid, fixed }) => {
-  const data = (fluid && fluid[0]) || (fixed && fixed[0])
+  const data = fluid ? getCurrentSrcData(fluid) : getCurrentSrcData(fixed)
 
   return data.src
+}
+
+/**
+ * Returns the current src - Preferably with art-direction support.
+ * @param currentData  {{media?: string}[]}   The fluid or fixed image array.
+ * @return {{src: string, media?: string}}
+ */
+const getCurrentSrcData = currentData => {
+  if (isBrowser && hasArtDirectionSupport(currentData)) {
+    // Do we have an image for the current Viewport?
+    const foundMedia = currentData.findIndex(matchesMedia)
+    if (foundMedia !== -1) {
+      return currentData[foundMedia]
+    }
+  }
+  // Else return the first image.
+  return currentData[0]
 }
 
 // Cache if we've seen an image before so we don't bother with
@@ -416,7 +452,7 @@ class Image extends React.Component {
 
     if (fluid) {
       const imageVariants = fluid
-      const image = imageVariants[0]
+      const image = getCurrentSrcData(fluid)
 
       return (
         <Tag
@@ -516,7 +552,7 @@ class Image extends React.Component {
 
     if (fixed) {
       const imageVariants = fixed
-      const image = imageVariants[0]
+      const image = getCurrentSrcData(fixed)
 
       const divStyle = {
         position: `relative`,