+++ www/src/components/code-block/index.js
@@ -2,10 +2,17 @@
 import { jsx } from "theme-ui"
 import React from "react"
 import PropTypes from "prop-types"
-import Highlight, { defaultProps } from "prism-react-renderer"
+import loadable from "@loadable/component"
 
 import Copy from "../copy"
 import normalize from "./normalize"
+
+const LazyHighlight = loadable(async () => {
+  const Module = await import(`prism-react-renderer`)
+  const Highlight = Module.default
+  const defaultProps = Module.defaultProps
+  return props => <Highlight {...defaultProps} {...props} />
+})
 
 const getParams = (name = ``) => {
   const [lang, params = ``] = name.split(`:`)
@@ -43,12 +50,7 @@ const CodeBlock = ({
   )
 
   return (
-    <Highlight
-      {...defaultProps}
-      code={content}
-      language={language}
-      theme={undefined}
-    >
+    <LazyHighlight code={content} language={language} theme={undefined}>
       {({ tokens, getLineProps, getTokenProps }) => (
         <React.Fragment>
           {title && (
@@ -95,7 +97,7 @@ const CodeBlock = ({
           </div>
         </React.Fragment>
       )}
-    </Highlight>
+    </LazyHighlight>
   )
 }